functions/requirements.txt
azure-functions>=1.18.0
azure-functions-durable>=1.2.9
httpx>=0.27.0
sqlalchemy>=2.0
psycopg2-binary>=2.9
onnxruntime>=1.17.0
numpy>=1.26
Pillow>=10.0
openai>=1.12.0
